Swap emojis used for Alfil and Ferz.
Alfil (2.2: jumps two diagonal spaces) now uses the bishop symbol, and
Ferz (1.1: moves on diagonal space) uses the pawn symbol.

This is more consistent with the expectation that a bishop moves further
than a pawn. I originally swapped the pieces because there are 8 Alfils
and only 2 Ferzes, and I was worried having 8 bishops would be too visually
distracting, but after playing a bit I don't find it to be a problem.